agenticone-backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
numpy==1.24.3
markdown2==2.5.4
reportlab==4.0.4
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.0.0
PyPDF2==3.0.1
python-docx==0.8.11